fix(order): 행배정 UPDATE의 $2 타입추론 실패 수정 (주문 시트 미반영 근본원인) (#149)
운영 진단에서 order_submissions가 written/queued 0, pending 5205·pending_no_row 280으로
전량 적체. 막힌 주문 sheet_error가 PostgreSQL "could not determine data type of
parameter $2". 행배정 트랜잭션의 UPDATE(`SET sheet_row=$2 ... CASE WHEN $2 IS NULL`)가
파라미터 $2 타입을 추론 못해 매 호출 throw → claim 롤백 → 주문이 시트에 못 들어감.
order-ledger(#135)가 이 버그로 사실상 깨져 있었고, reconcile도 같은 쿼리를 복사해
동일 버그였음(그래서 복구도 실패).

- createOrderLedgerEntry · reconcileStuckOrders의 UPDATE: $2 → $2::int 명시 캐스트로
  타입 추론 실패 제거(SET·CASE WHEN 양쪽).
- markOrderWritten의 COALESCE($2,...)도 $2::int로 예방 캐스트(쓰기 성공 경로).

검증: node --check 통과, orderLedger 단위테스트 통과.
배포 후 order-reconcile로 적체분(~5500건) 일괄 복구 예정.

+++ server/src/services/orderLedger.service.js
@@ -584,10 +584,10 @@ async function createOrderLedgerEntry(input) {
         });
         await client.query(
           `UPDATE order_submissions
-              SET sheet_row = $2,
+              SET sheet_row = $2::int,
                   tab_gid = COALESCE(NULLIF($3, ''), tab_gid),
-                  mirror_status = CASE WHEN $2 IS NULL THEN 'pending_no_row' ELSE 'pending' END,
-                  sheet_error = CASE WHEN $2 IS NULL THEN $4 ELSE NULL END
+                  mirror_status = CASE WHEN $2::int IS NULL THEN 'pending_no_row' ELSE 'pending' END,
+                  sheet_error = CASE WHEN $2::int IS NULL THEN $4 ELSE NULL END
             WHERE id = $1`,
           [
             orderSubmissionId,
@@ -653,7 +653,7 @@ async function markOrderWritten(orderSubmissionId, sheetRow) {
   await getPool().query(
     `UPDATE order_submissions
         SET mirror_status = 'written',
-            sheet_row = COALESCE($2, sheet_row),
+            sheet_row = COALESCE($2::int, sheet_row),
             sheet_written_at = NOW(),
             sheet_error = NULL
       WHERE id = $1`,
@@ -791,10 +791,10 @@ async function reconcileStuckOrders({ limit = 50, perTabCap = 20, sheetId = null
       });
       await client.query(
         `UPDATE order_submissions
-            SET sheet_row = $2,
+            SET sheet_row = $2::int,
                 tab_gid = COALESCE(NULLIF($3, ''), tab_gid),
-                mirror_status = CASE WHEN $2 IS NULL THEN 'pending_no_row' ELSE 'pending' END,
-                sheet_error = CASE WHEN $2 IS NULL THEN $4 ELSE NULL END
+                mirror_status = CASE WHEN $2::int IS NULL THEN 'pending_no_row' ELSE 'pending' END,
+                sheet_error = CASE WHEN $2::int IS NULL THEN $4 ELSE NULL END
           WHERE id = $1`,
         [row.id, claim.row || null, tabContext.tabGid || gid,
          claim.error || (claim.exhausted ? 'row claim exhausted' : 'row claim failed')]
